Fix I2C device ping bus lock up
Supports #292.

Successfully addressing a device for a read operation results in the
device taking control of the SDA signal. The device will retain control
of the SDA signal until a NACK terminated read is performed.

The ping function previously attempted to transmit a stop condition
after it had attempted to address a device. This caused the bus to lock
up when a device was successfully addressed for a read operation. The
pin function now performs a NACK terminated read when a device is
successfully addressed for a read operation to avoid bus lock up.